+++ server/src/main/java/org/apache/druid/server/coordinator/duty/BalanceSegments.java
@@ -251,14 +251,15 @@ public class BalanceSegments implements CoordinatorDuty
             if (moveSegment(segmentToMoveHolder, destinationHolder.getServer(), params)) {
               moved++;
             } else {
+              log.info("Segment [%s] cannot be moved. Maybe no space left or already being served or already in queue.", segmentToMove.getId());
               unmoved++;
             }
           } else {
-            log.debug("Segment [%s] is 'optimally' placed.", segmentToMove.getId());
+            log.info("Segment [%s] is 'optimally' placed.", segmentToMove.getId());
             unmoved++;
           }
         } else {
-          log.debug("No valid movement destinations for segment [%s].", segmentToMove.getId());
+          log.info("No valid movement destinations for segment [%s].", segmentToMove.getId());
           unmoved++;
         }
       }
@@ -291,7 +292,7 @@ public class BalanceSegments implements CoordinatorDuty
     if (!toPeon.getSegmentsToLoad().contains(segmentToMove) &&
         (toServer.getSegment(segmentId) == null) &&
         new ServerHolder(toServer, toPeon).getAvailableSize() > segmentToMove.getSize()) {
-      log.debug("Moving [%s] from [%s] to [%s]", segmentId, fromServer.getName(), toServer.getName());
+      log.info("Moving [%s] from [%s] to [%s]", segmentId, fromServer.getName(), toServer.getName());
 
       LoadPeonCallback callback = null;
       try {
